kubernetes: upgrade 1.18.8 -> 1.22.1
Upgrade kubernetes to latest stable release.

We drop our bindata patch, since the generated bindata is no longer
used. Also drop patches that were backported and are now part of the
releae. Drop patch to "fix arm node failed.." since it was added for the
stable recipe and not refernced in the upstream recipe.

Issue: LINCD-5447

(LOCAL REV: NOT UPSTREAM) -- WRLINUX SPECIFIC

Signed-off-by: Sakib Sajal <sakib.sajal@windriver.com>

Rebased on 20220105:
* dropped modification for kubernetes_git.bb that the dependency has
  been added by upstream
* squash 2 following local patches for kubernetes 1.22.1
  - a7bc8b3f kubernetes-1.22.1: drop arm specific patch
  - dd790862 kubernetes: update sed expression

Signed-off-by: Kai Kang <kai.kang@windriver.com>