kubernetes: fix install-vs-shipped qa issue with usrmerge
Issue: LINCD-11808

The feature/usrmerge is enabled, /bin/k8s-init is reported as installed
but not shipped. This is because the k8s recipe is missing the setting
of BIN_PREFIX. Set this variable to the default value as set in k3s recipe.

(LOCAL REV: NOT UPSTREAM) -- WRLinux Specific Patch

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>