docker-distribution: fix for CVE-2023-2253
Issue: LIN1023-520

CVE: CVE-2023-2253

A flaw was found in the `/v2/_catalog` endpoint in distribution/distribution,
which accepts a parameter to control the maximum number of records returned
(query string: `n`). This vulnerability allows a malicious user to submit an
unreasonably large value for `n,` causing the allocation of a massive string
array, possibly causing a denial of service through excessive use of memory.

References:
https://github.com/distribution/distribution/security/advisories/GHSA-hqxw-f8mx-cpmw
https://github.com/distribution/distribution/commit/521ea3d973cb0c7089ebbcdd4ccadc34be941f54

(LOCAL REV: NOT UPSTREAM) -- WRLinux Specific Patches

Signed-off-by: Narpat Mali <narpat.mali@windriver.com>