openvswitch: fix CVE-2023-3966
Issue: LIN1023-3637

A flaw was found in Open vSwitch where multiple versions are
vulnerable to crafted Geneve packets, which may result in a
denial of service and invalid memory accesses. Triggering
this issue requires that hardware offloading via the netlink
path is enabled.

Reference:
https://security-tracker.debian.org/tracker/CVE-2023-3966

(LOCAL REV: NOT UPSTREAM) -- Not applicable, upstream out of maintenance

Signed-off-by: Yogita Urade <yogita.urade@windriver.com>